// packages/astro/dist/types/public/config.d.ts
export interface AstroConfig extends AstroConfigType {
	integrations: AstroIntegration[];
}